+++ src/main/java/insane96mcp/progressivebosses/module/dragon/feature/LarvaFeature.java
@@ -5,7 +5,7 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.utils.RandomHelper;
 import insane96mcp.progressivebosses.base.Strings;
-import insane96mcp.progressivebosses.module.dragon.ai.dragon.PBNearestAttackableTargetGoal;
+import insane96mcp.progressivebosses.module.dragon.ai.PBNearestAttackableTargetGoal;
 import insane96mcp.progressivebosses.setup.Config;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ai.attributes.AttributeModifier;